requirements.txt
faiss-cpu
fastapi
uvicorn
sentence-transformers
faiss-cpu
chromadb
streamlit
requests
google-genai>=1.0.0
google-cloud-aiplatform
google-cloud-storage
langchain-community